ref
Корректировка папок. Создание класса ProductCategory

+++ demo/src/main/java/com/example/dto/OrderDetailDTO.java
@@ -1,5 +1,7 @@
 package com.example.dto;
 
+import java.math.BigDecimal;
+
 import com.example.enums.OrderStatus;
 
 import lombok.Data;
@@ -9,5 +11,5 @@ public class OrderDetailDTO {
 
     private int id;
     private OrderStatus orderStatus;
-    private double totalAmount;
+    private BigDecimal totalAmount;
 }
